Add @javax.inject.Inject support, revert gwt RequestResourceVisitor to RequestResourceBuilder and expose internal RequestResourceVisitor.uri method
Inject simplify configuration allowing bind(MyRestService.class).to(MyRestService_RestServiceModel.class).
RequestResourceBuilder implements visitor, but it's a builder and can be used as an stand alone builder.
And the uri and query methods exposed in RequestResourceBuilder should be considered internal, exposed is useful to see or use (hacking) the generated uri to open it directly.