Improve RequestResourceBuilder so dispatcher can detect cancellations
This is useful if you use the dispatcher to show a loading indicator, before this change, canceled request aren't notified to the dispacher, now, a CancelRequestException is passed to the onError handler so the loading indicater can be turned off. The examplo project have been improved to show an example usage of status indicator with cancelling support.