Apply encodeComponent to each path and move baseUrl as independent field
The baseUrl is never encoded, always passthroughs but now it must be defined on construction. This simplify paths logic, and make it more predictable and customizable. The are some big changes in behaviour, although generated code wont get affected. Previously, you can call .path("a/b/c") and it add it as a actual path, now it gets encoded so the whole "a/b/c" is just one component "a%2Fb%2Fc". The processor always split the path and generate a call like .path("a", "b", "c") so it still works correctly but if you are using the collector manually it might break your code, sorry.